renamed and fixed FileDownloadTest,FileUploadTest; fixed BaseTest
